requirements.txt
funcsigs==1.0.2
mock==2.0.0
numpy==1.11.3
pbr==1.10.0
protobuf==3.1.0.post1
python-speech-features==0.4
scikits.audiolab==0.11.0
scipy==0.18.1
six==1.10.0
tensorflow-gpu==0.12.1
tqdm==4.10.0
